scripts/requirements.txt
python-dotenv
docling
langchain
langchain-core
langchain-community
langchain-openai
chromadb
